requirements.txt
autopep8==1.4.3
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.13.0
coreapi==2.3.3
coreschema==0.0.4
cssutils==1.0.2
defusedxml==0.5.0
Django==2.1.7
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-filter==2.0.0
django-geoposition==0.3.0
django-inlinecss==0.1.2
django-openid-auth==0.15
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Faker==0.9.1
flake8==3.5.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
openapi-codec==1.3.2
pbr==5.1.3
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pynliner3==0.6
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.8
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1
yapf==0.26.0
